Instrument each WebRTC Track
This will create a new Sentry transaction each time metrics are fetched
(not ideal), but we can likely sample these. I tried to stick to the
metrics that I thought would matter the most, since we can only collect
10 at a go here.

I think in the future this may be better off as a promfile that we throw
over the WebSocket or something? POST it? Have the server (or api-deux?)
request (and intercept) it to export to prom?

Signed-off-by: Paul Tagliamonte <paul@kittycad.io>