Retry when the Server flakes out on us.
This isn't ideal, and I'd love to remove this eventually; but this is
helpful when we exhaust the Engine pool, or when running locally. This
will retry a few times, and only fail if there's a few repeated
failures.

Since the WebSocket fully handshakes, and then fails after handshake
(when api-deux realizes we're out of Engines, specifically), I can't
check the HTTP code directly where I wanted to.

This is kludgey for now, and it may mask flaky endpoints; but it may be
worth it in the short-term. I'd love to find a way for the modeling
endpoint to not upgrade via api-deux unless there's an engine for it
long-er term.