KCL: Refactor pipe bodies
PipeExpressions should always have at least two child expressions. The
first expression (the head) is a little different from the following
subexpressions, because the head can be any value. But the remaining
expressions (tail) must be function calls.

This refactors the AST node for PipeExpression so that information is
encoded in the type system, not just checked at runtime. Instead of a
Vec<Value>, it should have a separate head Value and tail Vec<Value>.