Add in a Fuzzer for KCL's parser and lexer
This found https://github.com/KittyCAD/modeling-app/pull/2491 and I want
to chip away at covering more of the input parsers with a fuzzer as time
goes on - this was just the easiest one to do first. I'm thinking of
using something a bit more structured long-term, but for now this'll do.

Signed-off-by: Paul R. Tagliamonte <paul@kittycad.io>