patternTransform KCL stdlib call
Fixes #2735

Adds a new KCL stdlib function pattern(numRepetitions, transform, whatToRepeat) which replicates some geometry the given number of times, applying the given transform to each repetition.

This is the the KCL stdlib function which accepts another function as input, so there's some small refactors to make that possible.

Still TODO:

Generate a KCL docs schema for the transform parameter (we've never had to generate docs for higher-order functions before)