Update playwright test
The test was failing because 'myVar' could be found in two places, and
it was told to just "click on 'myVar'". That wasn't ambiguous before when
the code editor had `const myVar =` but now with the naked assignment, it is.

Solution is to make it more specific and find 'myVar' in the code editor.