Remove duplicate JSON "type" tags
KCL serializes its AST nodes serialize to JSON. This way they can be read by the frontend. We also serialize to JSON for the parser's snapshot tests.

There's an unfortunate effect where we use `serde(tag = "type")` on both the `enum Expr` and on the individual AST node `struct` types stored within that enum. So the `Expr::Literal(Literal{...})` gets _two_ `type: "Literal"` tags (one from the enum wrapper, and one from the struct). This just seems weird, you get JSON like `{"type": "Literal", "type": "Literal", value: 3}`. Duplicate keys are technically OK JSON, but it's weird and adds a lot of noise to our snapshot tests.

We can't remove the tag from the individual structs, because the frontend needs to check (at runtime) what kind of node each AST node is. So we have to remove it from the enum. But there's a problem: `#[serde(tag)]` on a struct is only used when serializing. It doesn't get checked when deserializing. So we cannot rely on that tag attribute to deserialize `enum Expr` into the correct variants.

So, we have to keep the tag on each struct, but make sure it's read during deserialization.

Define a new tag type for `Literal` which serializes to the string `"Literal"`. It becomes part of the `struct Literal`. We do this for each kind of expression struct, then change the `serde(tag = "type")`  on `enum Expr` to `serde(untagged)`.

You can look at the snapshot tests and see a lot of removed JSON duplicate keys. I also added a new test in `ast/types.rs` showing that serde can now serialize then deserialize an AST. Previously this test would have panicked, complaining about duplicate keys.