Path enum no longer includes BasePath
BasePath is a pair of points and geometry metadata.

Previously BasePath was used for:

 - The start of a sketch
 - The data which all path types have in common

It wasn't a good fit for the start of a sketch, because there was a lot
of duplicated information, and a sketch starts at a point, not a line
(two points).

This adds a separate SketchStart struct which replaces the first use
of BasePath. This means BasePath no longer needs to be a variant of
Path, simplifying it.

Solves problem 4 of
 https://github.com/KittyCAD/modeling-app/issues/4297