KCL: Executor returns specific errors, not anyhow
Some clients (e.g. the simulation tests) want to know
why a KCL program failed. Was it because of a network error?
Or because the client wasn't authorized properly? Or
was it a KCL runtime type error?

It's difficult for clients to ask these questions,
because the interpreter just returns Anyhow error,
basically just a string.

Instead, we should return different variants of an Error
enum for different kinds of errors.

This lets us render nice Cargo-style error messages
that show the exact KCL line where an error occurred,
instead of just printing off a source range.